Refactor KnowledgeModule into smaller tab subcomponents
Extracted the monolithic KnowledgeModule component into separate subcomponents:
- KnowledgeVaultTab
- IngestionTab
- ScholarChatTab
- SourcesTab

Used CSS display toggling to preserve the internal state (e.g., chat history, upload progress) of these components when switching tabs.
